+++ crates/miden-protocol/src/transaction/kernel/advice_inputs.rs
@@ -350,18 +350,18 @@ impl TransactionAdviceInputs {
             let note_arg = tx_inputs.tx_args().get_note_args(note.id()).unwrap_or(&EMPTY_WORD);
 
             // recipient inputs / assets commitments
-            self.add_map_entry(recipient.inputs().commitment(), recipient.inputs().to_elements());
+            self.add_map_entry(recipient.storage().commitment(), recipient.storage().to_elements());
             self.add_map_entry(assets.commitment(), assets.to_padded_assets());
 
             // note details / metadata
             note_data.extend(recipient.serial_num());
             note_data.extend(*recipient.script().root());
-            note_data.extend(*recipient.inputs().commitment());
+            note_data.extend(*recipient.storage().commitment());
             note_data.extend(*assets.commitment());
             note_data.extend(*note_arg);
             note_data.extend(note.metadata().to_header_word());
             note_data.extend(note.metadata().to_attachment_word());
-            note_data.push(recipient.inputs().num_values().into());
+            note_data.push(recipient.storage().len().into());
             note_data.push((assets.num_assets() as u32).into());
             note_data.extend(assets.to_padded_assets());