+++ crates/miden-testing/src/kernel_tests/tx/test_prologue.rs
@@ -485,7 +485,7 @@ fn input_notes_memory_assertions(
 
         assert_eq!(
             exec_output.get_note_mem_word(note_idx, INPUT_NOTE_INPUTS_COMMITMENT_OFFSET),
-            note.inputs().commitment(),
+            note.storage().commitment(),
             "note input commitment should be stored at the correct offset"
         );